Implemented usage of repository in ProductService

+++ backend/src/main/java/com/inventory_system/backend/controller/ProductController.java
@@ -23,7 +23,7 @@ public class ProductController {
 
     // Build Get Products REST API
     @GetMapping("{id}")
-    public ResponseEntity<ProductDTO> getProductById (@PathVariable("id") int productId) {
+    public ResponseEntity<ProductDTO> getProductById (@PathVariable("id") Long productId) {
         ProductDTO productDTO = productService.getProductById(productId);
         return ResponseEntity.ok(productDTO);
     }
@@ -37,7 +37,7 @@ public class ProductController {
 
     // Build Update Products REST API
     @PutMapping("/{id}")
-    public ResponseEntity<ProductDTO> updateProduct(@PathVariable("id") int productId,
+    public ResponseEntity<ProductDTO> updateProduct(@PathVariable("id") Long productId,
                                                     @RequestBody ProductDTO updatedProduct) {
         ProductDTO productDto = productService.updateProduct(productId, updatedProduct);
         return ResponseEntity.ok(productDto);
@@ -45,7 +45,7 @@ public class ProductController {
 
     // Build Delete Product REST API
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteProduct(@PathVariable("id") int productId) {
+    public ResponseEntity<String> deleteProduct(@PathVariable("id") Long productId) {
         productService.deleteProduct(productId);
         return ResponseEntity.ok("Product deleted successfully");
     }
